bsc-analyzer/requirements.txt
# Lumina BSC Analyzer - Python Dependencies
# ==========================================

# Core async
aiohttp>=3.9.0
websockets>=12.0

# Web3
web3>=6.0.0

# Database
clickhouse-driver>=0.2.6

# ML (KOL scorer: Phase 2 train + Phase 3 backtest)
lightgbm>=4.0.0
treelite>=4.0.0
# Legacy / draft experiments
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
shap>=0.44.0  # Optional: for feature importance analysis

# Visualization (optional, for training)
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0
kaleido>=0.2.1  # Plotly static export

# Notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Parquet output
pyarrow>=14.0.0

# Utilities
python-dotenv>=1.0.0
